use NonZero types

+++ src/ffi/totp_c_binds.rs
@@ -1,6 +1,5 @@
 use crate::ffi::converter::{to_str, to_totp};
-use std::ffi::c_ulonglong;
-use std::os::raw::{c_char, c_ushort};
+use std::ffi::{c_char, c_longlong, c_ulonglong, c_ushort};
 
 /// Configuration for TOTP (Time-based One-Time Password).
 ///
@@ -17,7 +16,7 @@ pub struct TotpConfig {
     pub secret: *const c_char,
     pub length: c_ushort,
     pub radix: c_ushort,
-    pub interval: c_ushort,
+    pub interval: c_longlong,
 }
 
 /// Generates a TOTP (Time-based One-Time Password) based on the provided configuration for the current time.
